Add script to resize already annotated images on a Label Studio project
The intended usage of the script is to save time when there is a previously created and annotated Label Studio project, and you want to resize all image KEEPING the existing annotations.

If you attempt to do it manually the problem is that annotations are not scaled according to the new size of the images.

A `/scripts` folder was created with the new script and a requirements file.