Download only the labeled dataset images
Now only the labeled images are downloaded from the bucket, instead of downloading ALL images and then filtering the labeled ones. This is faster when there are many images on the bucket.

Some `os` calls were also replaced by calls to libraries, which is faster and less prone to errors. Specifically for decompressing the YOLO zip file, and for getting the label config from LabelStudio.

Methods were also reordered and categorized.